+++ src/main/java/codingforlove/community/Service/QuestionService.java
@@ -1,17 +1,16 @@
 package codingforlove.community.Service;
 
 import codingforlove.community.DTO.PaginationDTO;
-import codingforlove.community.DTO.QuestionDTO;
 import org.springframework.ui.Model;
 
 public interface QuestionService {
     PaginationDTO list(Model model, Integer page, Integer size);
 
-    PaginationDTO list(int id, Integer page, Integer size);
+    PaginationDTO list(Long id, Integer page, Integer size);
 
-    String getById(Integer id, Model model);
+    String getById(Long id, Model model);
 
-    void incView(Integer id);
+    void incView(Long id);
 
-    String deleteById(Integer id);
+    String deleteById(Long id);
 }
